# [Enhancement] Improve performance of `strings::memcpy_inlined`