@dirtysalt 2022-11-17
run starrocks_admit_test